requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.10.0,<2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
pypdf>=3.17.4
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
